requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
anthropic
openai
psycopg2-binary
